Split dev pages in two different ones
